requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.11.0=pypi_0
addict=2.4.0=pypi_0
argon2-cffi=20.1.0=py38h27cfd23_1
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.1=pyhd3eb1b0_0
brotli=1.0.9=he6710b0_2
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.1=pypi_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chamferdist=1.0.0=pypi_0
chardet=4.0.0=pypi_0
cloudpickle=1.6.0=py_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
cython-bbox=0.1.3=pypi_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2021.8.0=pyhd3eb1b0_0
dataclasses=0.6=pypi_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
detectron2=0.3=dev_0
entrypoints=0.3=py38_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd3eb1b0_0
future=0.18.2=pypi_0
google-auth=1.25.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
grpcio=1.35.0=pypi_0
h5py=3.6.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
jedi=0.17.0=py38_0
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab-widgets=1.0.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
locket=0.2.1=py38h06a4308_1
lz4-c=1.9.3=h2531618_0
markdown=3.3.3=pypi_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.3.4=pypi_0
matplotlib-base=3.4.2=py38hab158f2_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=py_0
nb_conda=2.2.1=py38_1
nb_conda_kernels=2.3.1=py38h06a4308_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=pypi_0
ninja=1.10.2=pypi_0
notebook=6.4.0=py38h06a4308_0
numpy=1.20.0=pypi_0
numpy-base=1.20.3=py38h74d4b33_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
open3d=0.12.0=pypi_0
opencv-python=4.5.4.58=pypi_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.2.2=pypi_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py38h2c7a002_0
pip=21.1.3=py38h06a4308_0
plyfile=0.7.3=pypi_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.14.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pydot=1.4.1=pypi_0
pyellipsoid=0.0.3=pypi_0
pygments=2.9.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pypng=0.0.20=pypi_0
pyrsistent=0.17.3=py38h7b6447c_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.0=py3.8_cuda11.0.221_cudnn8.0.3_0
pytorch3d=0.3.0=pypi_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=5.4.1=py38h27cfd23_1
pyzmq=20.0.0=py38h2531618_1
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7=pypi_0
scikit-image=0.18.1=py38ha9443f7_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.0=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sklearn=0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.7=py38_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.10.1=py38hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torch=1.10.1=pypi_0
torchaudio=0.7.2=pypi_0
torchvision=0.11.2=pypi_0
tornado=6.1=py38h27cfd23_0
traitlets=5.0.5=pyhd3eb1b0_0
trimesh=3.9.8=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.3=pypi_0
visualpriors=0.3.5=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
